+++ R/pat_sample.R
@@ -202,7 +202,7 @@ pat_sample <- function(
     dplyr::distinct() %>% 
     dplyr::arrange(.data$datetime)
   
-  # ----- Create the Purple Air Timeseries (pat) object ------------------------
+  # ----- Create the PurpleAir Timeseries (pat) object ------------------------
   
   # Combine meta and data dataframes into a list
   pat <- list(meta = pat$meta, data = data)