+++ local_jon/test_archiveBaseDir.R
@@ -9,10 +9,10 @@ setArchiveBaseDir("/Users/jonathan/Projects/PWFSL/2020/sensor-data-ingest-v1/out
 
 pas <- pas_load()
 
-pat <- pat_createNew(pas, label = "Ballard")
+pat <- pat_createNew(label = "Ballard", pas = pas)
 pat_multiplot(pat)
 
-as <- pat_createAirSensor(pat)
-PWFSLSmoke::monitor_timeseriesPlot(as, shadedNight = TRUE)
+sensor <- pat_createAirSensor(pat)
+PWFSLSmoke::monitor_timeseriesPlot(sensor, shadedNight = TRUE)